Fixed pr issue Moved Popup examples to another dir
